% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkimball.R
\name{dkimball}
\alias{dkimball}
\title{Returns the cumulative distribution function}
\usage{
dkimball(x, S, L, min = 0)
}
\arguments{
\item{x}{A positive numerical input vector requiring evaluation}

\item{S}{A parameter for the maximum of the truncated normal distribution. Must be positive.}

\item{L}{Mean parameter. Must be positive.}

\item{min}{A parameter for the minimum of the truncated normal distribution. Must be positive. Default is zero.}
}
\value{
A numerical report of the cumulative probability of x
}
\description{
Returns the cumulative distribution function
}
\examples{
dkimball(3, 8, 4, min=0)
}
