% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rkimball.R
\name{rkimball}
\alias{rkimball}
\title{Returns random deviates from the Kimball distribution}
\usage{
rkimball(x, S, L, min = 0)
}
\arguments{
\item{x}{A value describing the desired number of random deviates (length of a vector)}

\item{S}{A parameter for the maximum of the truncated normal distribution. Must be positive.}

\item{L}{Mean parameter. Must be positive.}

\item{min}{A parameter for the minimum of the truncated normal distribution. Must be positive. Default is zero.}
}
\value{
A vector or random deviates
}
\description{
Returns random deviates from the Kimball distribution
}
\examples{
rkimball(100, 8, 4, min=0)
}
